clean up ens deployment
