+++ packages/contracts/deploy/05_ens.ts
@@ -1,82 +1,27 @@
 import {HardhatRuntimeEnvironment} from 'hardhat/types';
 import {DeployFunction} from 'hardhat-deploy/types';
-import {getContractAddress} from './helpers';
+import {getContractAddress, setupENS} from './helpers';
 
-function ensLabelHash(label: string, ethers: any): string {
-  return ethers.utils.id(label);
-}
-
-function ensDomainHash(name: string, ethers: any): string {
-  return ethers.utils.namehash(name);
-}
+const ensAddresses: {[key: string]: string} = {
+  mainnet: '0x314159265dD8dbb310642f98f50C066173C1259b',
+  ropsten: '0x112234455C3a32FD11230C42E7Bccd4A84e02010',
+  rinkeby: '0xe7410170f87102DF0055eB195163A03B7F2Bff4A',
+  goerli: '0x112234455C3a32FD11230C42E7Bccd4A84e02010',
+};
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
-  const {deployments, getNamedAccounts, ethers} = hre;
+  const {deployments, getNamedAccounts, ethers, getChainId, network} = hre;
   const {deploy} = deployments;
 
   const {deployer} = await getNamedAccounts();
 
-  const ensRet = await deploy('ENSRegistry', {
-    from: deployer,
-    log: true,
-  });
-  const ensRegistryAddress: string = ensRet.receipt?.contractAddress || '';
-
-  const ensResolverRet = await deploy('PublicResolver', {
-    from: deployer,
-    args: [ensRegistryAddress, ethers.constants.AddressZero],
-    log: true,
-  });
-  const ensResolverAddress: string =
-    ensResolverRet.receipt?.contractAddress || '';
-
-  // setup resolver
-  const ensRegistryContract = await ethers.getContractAt(
-    'ENSRegistry',
-    ensRegistryAddress
-  );
-  const ensResolverContract = await ethers.getContractAt(
-    'PublicResolver',
-    ensResolverAddress
-  );
-  await ensRegistryContract.setSubnodeOwner(
-    ensDomainHash('', ethers),
-    ensLabelHash('resolver', ethers),
-    deployer
-  );
-
-  const resolverNode = ensDomainHash('resolver', ethers);
-
-  await ensRegistryContract.setResolver(resolverNode, ensResolverAddress);
-  await ensResolverContract['setAddr(bytes32,address)'](
-    resolverNode,
-    ensResolverAddress
-  );
-
-  await ensRegistryContract.setSubnodeOwner(
-    ensDomainHash('', ethers),
-    ensLabelHash('eth', ethers),
-    deployer
-  );
-
-  await ensRegistryContract.setSubnodeOwner(
-    ensDomainHash('eth', ethers),
-    ensLabelHash('dao', ethers),
-    deployer
-  );
-
   const managingDAOAddress = await getContractAddress('DAO', hre);
   const node = ethers.utils.namehash('dao.eth');
 
-  // deterministic
-  const [owner] = await ethers.getSigners();
-  const nonce = await owner.getTransactionCount();
-  const futureAddress = ethers.utils.getContractAddress({
-    from: deployer,
-    nonce: nonce + 2, // next address is implementation, so we use +2 for the proxy address
-  });
-
-  await ensRegistryContract.setApprovalForAll(futureAddress, true);
+  let ensRegistryAddress = ensAddresses[network.name];
+  if (!ensRegistryAddress) {
+    ensRegistryAddress = await setupENS(hre);
+  }
 
   await deploy('ENSSubdomainRegistrar', {
     from: deployer,
